moved navbar and mobilescreen hook to index
