+++ src/pages/index.js
@@ -1,8 +1,10 @@
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
+import { useState, useEffect } from "react";
 import Landing from "./Landing";
 import AboutUs from "../components/AboutUs";
 import DesignerFriendly from "../components/DesignerFriendly";
 import Register from "../components/Register";
+import NavBar from "../components/NavBar";
 
 export async function getStaticProps({ locale }) {
   return {
@@ -14,8 +16,24 @@ export async function getStaticProps({ locale }) {
 }
 
 export default function Home() {
+  
+  const [isMobileScreen, setIsMobileScreen] = useState(false);
+
+  useEffect(() => {
+        const handleResize = () => {
+            setIsMobileScreen(window.innerWidth < 768);
+        };
+        handleResize();
+        window.addEventListener('resize', handleResize);
+        return () => {
+            window.removeEventListener('resize', handleResize);
+        };
+  }, []);
+
+
   return (
     <div className="[background:linear-gradient(180deg,_#241b38,_#231f20)] grid grid-cols-[1fr_min(115ch,_calc(100%-64px))_1fr] gap-x-[32px] [&>*]:col-span-1 [&>*]:col-start-2 desktop:gap-x-[7.5vh] mobile:gap-x-[5vh] place-content-center">
+      <NavBar isMobileScreen={isMobileScreen}/>
       <Landing />
       <AboutUs />
       <DesignerFriendly />
